+++ codec.pb.go
@@ -71,25 +71,23 @@ func (m *Metadata) GetSchema() uint32 {
 	return 0
 }
 
-// CommitInfo is basically a replica of tendermint commit info
-// that allows us to store who signed the block in context.
-type CommitInfo struct {
-	Round int32      `protobuf:"varint,1,opt,name=round,proto3" json:"round,omitempty"`
-	Votes []VoteInfo `protobuf:"bytes,2,rep,name=votes,proto3" json:"votes"`
+// ValidatorUpdates represents latest validator state, currently used to validate SetValidatorMsg transactions.
+type ValidatorUpdates struct {
+	ValidatorUpdates []ValidatorUpdate `protobuf:"bytes,1,rep,name=validator_updates,json=validatorUpdates,proto3" json:"validator_updates"`
 }
 
-func (m *CommitInfo) Reset()         { *m = CommitInfo{} }
-func (m *CommitInfo) String() string { return proto.CompactTextString(m) }
-func (*CommitInfo) ProtoMessage()    {}
-func (*CommitInfo) Descriptor() ([]byte, []int) {
+func (m *ValidatorUpdates) Reset()         { *m = ValidatorUpdates{} }
+func (m *ValidatorUpdates) String() string { return proto.CompactTextString(m) }
+func (*ValidatorUpdates) ProtoMessage()    {}
+func (*ValidatorUpdates) Descriptor() ([]byte, []int) {
 	return fileDescriptor_9610d574777ab505, []int{1}
 }
-func (m *CommitInfo) XXX_Unmarshal(b []byte) error {
+func (m *ValidatorUpdates) XXX_Unmarshal(b []byte) error {
 	return m.Unmarshal(b)
 }
-func (m *CommitInfo) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
+func (m *ValidatorUpdates) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
 	if deterministic {
-		return xxx_messageInfo_CommitInfo.Marshal(b, m, deterministic)
+		return xxx_messageInfo_ValidatorUpdates.Marshal(b, m, deterministic)
 	} else {
 		b = b[:cap(b)]
 		n, err := m.MarshalTo(b)
@@ -99,136 +97,23 @@ func (m *CommitInfo) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
 		return b[:n], nil
 	}
 }
-func (m *CommitInfo) XXX_Merge(src proto.Message) {
-	xxx_messageInfo_CommitInfo.Merge(m, src)
+func (m *ValidatorUpdates) XXX_Merge(src proto.Message) {
+	xxx_messageInfo_ValidatorUpdates.Merge(m, src)
 }
-func (m *CommitInfo) XXX_Size() int {
+func (m *ValidatorUpdates) XXX_Size() int {
 	return m.Size()
 }
-func (m *CommitInfo) XXX_DiscardUnknown() {
-	xxx_messageInfo_CommitInfo.DiscardUnknown(m)
+func (m *ValidatorUpdates) XXX_DiscardUnknown() {
+	xxx_messageInfo_ValidatorUpdates.DiscardUnknown(m)
 }
 
-var xxx_messageInfo_CommitInfo proto.InternalMessageInfo
+var xxx_messageInfo_ValidatorUpdates proto.InternalMessageInfo
 
-func (m *CommitInfo) GetRound() int32 {
+func (m *ValidatorUpdates) GetValidatorUpdates() []ValidatorUpdate {
 	if m != nil {
-		return m.Round
-	}
-	return 0
-}
-
-func (m *CommitInfo) GetVotes() []VoteInfo {
-	if m != nil {
-		return m.Votes
+		return m.ValidatorUpdates
 	}
 	return nil
-}
-
-// VoteInfo tells us whether a validator signed the last block or not.
-type VoteInfo struct {
-	Validator       Validator `protobuf:"bytes,1,opt,name=validator,proto3" json:"validator"`
-	SignedLastBlock bool      `protobuf:"varint,2,opt,name=signed_last_block,json=signedLastBlock,proto3" json:"signed_last_block,omitempty"`
-}
-
-func (m *VoteInfo) Reset()         { *m = VoteInfo{} }
-func (m *VoteInfo) String() string { return proto.CompactTextString(m) }
-func (*VoteInfo) ProtoMessage()    {}
-func (*VoteInfo) Descriptor() ([]byte, []int) {
-	return fileDescriptor_9610d574777ab505, []int{2}
-}
-func (m *VoteInfo) XXX_Unmarshal(b []byte) error {
-	return m.Unmarshal(b)
-}
-func (m *VoteInfo) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
-	if deterministic {
-		return xxx_messageInfo_VoteInfo.Marshal(b, m, deterministic)
-	} else {
-		b = b[:cap(b)]
-		n, err := m.MarshalTo(b)
-		if err != nil {
-			return nil, err
-		}
-		return b[:n], nil
-	}
-}
-func (m *VoteInfo) XXX_Merge(src proto.Message) {
-	xxx_messageInfo_VoteInfo.Merge(m, src)
-}
-func (m *VoteInfo) XXX_Size() int {
-	return m.Size()
-}
-func (m *VoteInfo) XXX_DiscardUnknown() {
-	xxx_messageInfo_VoteInfo.DiscardUnknown(m)
-}
-
-var xxx_messageInfo_VoteInfo proto.InternalMessageInfo
-
-func (m *VoteInfo) GetValidator() Validator {
-	if m != nil {
-		return m.Validator
-	}
-	return Validator{}
-}
-
-func (m *VoteInfo) GetSignedLastBlock() bool {
-	if m != nil {
-		return m.SignedLastBlock
-	}
-	return false
-}
-
-// Validator represents a validator;
-type Validator struct {
-	Address []byte `protobuf:"bytes,1,opt,name=address,proto3" json:"address,omitempty"`
-	Power   int64  `protobuf:"varint,3,opt,name=power,proto3" json:"power,omitempty"`
-}
-
-func (m *Validator) Reset()         { *m = Validator{} }
-func (m *Validator) String() string { return proto.CompactTextString(m) }
-func (*Validator) ProtoMessage()    {}
-func (*Validator) Descriptor() ([]byte, []int) {
-	return fileDescriptor_9610d574777ab505, []int{3}
-}
-func (m *Validator) XXX_Unmarshal(b []byte) error {
-	return m.Unmarshal(b)
-}
-func (m *Validator) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
-	if deterministic {
-		return xxx_messageInfo_Validator.Marshal(b, m, deterministic)
-	} else {
-		b = b[:cap(b)]
-		n, err := m.MarshalTo(b)
-		if err != nil {
-			return nil, err
-		}
-		return b[:n], nil
-	}
-}
-func (m *Validator) XXX_Merge(src proto.Message) {
-	xxx_messageInfo_Validator.Merge(m, src)
-}
-func (m *Validator) XXX_Size() int {
-	return m.Size()
-}
-func (m *Validator) XXX_DiscardUnknown() {
-	xxx_messageInfo_Validator.DiscardUnknown(m)
-}
-
-var xxx_messageInfo_Validator proto.InternalMessageInfo
-
-func (m *Validator) GetAddress() []byte {
-	if m != nil {
-		return m.Address
-	}
-	return nil
-}
-
-func (m *Validator) GetPower() int64 {
-	if m != nil {
-		return m.Power
-	}
-	return 0
 }
 
 // ValidatorUpdate represents an update to validator set.
@@ -241,7 +126,7 @@ func (m *ValidatorUpdate) Reset()         { *m = ValidatorUpdate{} }
 func (m *ValidatorUpdate) String() string { return proto.CompactTextString(m) }
 func (*ValidatorUpdate) ProtoMessage()    {}
 func (*ValidatorUpdate) Descriptor() ([]byte, []int) {
-	return fileDescriptor_9610d574777ab505, []int{4}
+	return fileDescriptor_9610d574777ab505, []int{2}
 }
 func (m *ValidatorUpdate) XXX_Unmarshal(b []byte) error {
 	return m.Unmarshal(b)
@@ -294,7 +179,7 @@ func (m *PubKey) Reset()         { *m = PubKey{} }
 func (m *PubKey) String() string { return proto.CompactTextString(m) }
 func (*PubKey) ProtoMessage()    {}
 func (*PubKey) Descriptor() ([]byte, []int) {
-	return fileDescriptor_9610d574777ab505, []int{5}
+	return fileDescriptor_9610d574777ab505, []int{3}
 }
 func (m *PubKey) XXX_Unmarshal(b []byte) error {
 	return m.Unmarshal(b)
@@ -339,9 +224,7 @@ func (m *PubKey) GetData() []byte {
 
 func init() {
 	proto.RegisterType((*Metadata)(nil), "weave.Metadata")
-	proto.RegisterType((*CommitInfo)(nil), "weave.CommitInfo")
-	proto.RegisterType((*VoteInfo)(nil), "weave.VoteInfo")
-	proto.RegisterType((*Validator)(nil), "weave.Validator")
+	proto.RegisterType((*ValidatorUpdates)(nil), "weave.ValidatorUpdates")
 	proto.RegisterType((*ValidatorUpdate)(nil), "weave.ValidatorUpdate")
 	proto.RegisterType((*PubKey)(nil), "weave.PubKey")
 }
@@ -349,29 +232,24 @@ func init() {
 func init() { proto.RegisterFile("codec.proto", fileDescriptor_9610d574777ab505) }
 
 var fileDescriptor_9610d574777ab505 = []byte{
-	// 352 bytes of a gzipped FileDescriptorProto
-	0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0xff, 0x4c, 0x51, 0xc1, 0x4e, 0xea, 0x40,
-	0x14, 0x6d, 0x81, 0x16, 0xb8, 0x40, 0x78, 0x6f, 0x42, 0x5e, 0x9a, 0xb7, 0xe8, 0x6b, 0xba, 0x6a,
-	0x9e, 0x06, 0x0d, 0xba, 0x73, 0x87, 0x2b, 0xa3, 0x46, 0x33, 0x09, 0x6e, 0xc9, 0xb4, 0x73, 0x45,
-	0x42, 0x61, 0x9a, 0xce, 0x00, 0xe1, 0x2f, 0xfc, 0x2c, 0x96, 0x2c, 0x5d, 0x19, 0x03, 0x3f, 0x62,
-	0x3a, 0x6d, 0xc1, 0xdd, 0x3d, 0xe7, 0x9e, 0x73, 0xe6, 0x4c, 0x2e, 0xb4, 0x22, 0xc1, 0x31, 0xea,
-	0x27, 0xa9, 0x50, 0x82, 0x58, 0x6b, 0x64, 0x2b, 0xfc, 0xdb, 0x9b, 0x88, 0x89, 0xd0, 0xcc, 0x45,
-	0x36, 0xe5, 0x4b, 0xdf, 0x87, 0xc6, 0x23, 0x2a, 0xc6, 0x99, 0x62, 0xe4, 0x0f, 0xd8, 0x32, 0x7a,
-	0xc3, 0x39, 0x73, 0x4c, 0xcf, 0x0c, 0x3a, 0xb4, 0x40, 0xfe, 0x13, 0xc0, 0xad, 0x98, 0xcf, 0xa7,
-	0xea, 0x6e, 0xf1, 0x2a, 0x48, 0x0f, 0xac, 0x54, 0x2c, 0x17, 0x5c, 0x8b, 0x2c, 0x9a, 0x03, 0x72,
-	0x06, 0xd6, 0x4a, 0x28, 0x94, 0x4e, 0xc5, 0xab, 0x06, 0xad, 0x41, 0xb7, 0xaf, 0x1f, 0xed, 0xbf,
-	0x08, 0x85, 0x99, 0x6b, 0x58, 0xdb, 0x7e, 0xfe, 0x33, 0x68, 0xae, 0xf1, 0x63, 0x68, 0x94, 0x0b,
-	0x72, 0x0d, 0xcd, 0x15, 0x8b, 0xa7, 0x9c, 0x29, 0x91, 0xea, 0xc8, 0xd6, 0xe0, 0x57, 0x69, 0x2e,
-	0xf9, 0xc2, 0x7d, 0x12, 0x92, 0xff, 0xf0, 0x5b, 0x4e, 0x27, 0x0b, 0xe4, 0xe3, 0x98, 0x49, 0x35,
-	0x0e, 0x63, 0x11, 0xcd, 0x9c, 0x8a, 0x67, 0x06, 0x0d, 0xda, 0xcd, 0x17, 0x0f, 0x4c, 0xaa, 0x61,
-	0x46, 0xfb, 0x37, 0xd0, 0x3c, 0x26, 0x11, 0x07, 0xea, 0x8c, 0xf3, 0x14, 0xa5, 0xd4, 0x8f, 0xb5,
-	0x69, 0x09, 0xb3, 0x7f, 0x25, 0x62, 0x8d, 0xa9, 0x53, 0xf5, 0xcc, 0xa0, 0x4a, 0x73, 0xe0, 0x8f,
-	0xa0, 0x7b, 0x34, 0x8f, 0x12, 0xce, 0x14, 0x92, 0x73, 0xa8, 0x27, 0xcb, 0x70, 0x3c, 0xc3, 0x4d,
-	0xd1, 0xb7, 0x53, 0xf4, 0x7d, 0x5e, 0x86, 0xf7, 0xb8, 0x29, 0xca, 0xda, 0x89, 0x46, 0xa7, 0xd8,
-	0xca, 0xcf, 0xd8, 0x4b, 0xb0, 0x73, 0x35, 0x21, 0x50, 0x53, 0x9b, 0x04, 0x75, 0x54, 0x93, 0xea,
-	0x39, 0xe3, 0xb2, 0x83, 0x68, 0x4b, 0x9b, 0xea, 0x79, 0xe8, 0x6c, 0xf7, 0xae, 0xb9, 0xdb, 0xbb,
-	0xe6, 0xd7, 0xde, 0x35, 0xdf, 0x0f, 0xae, 0xb1, 0x3b, 0xb8, 0xc6, 0xc7, 0xc1, 0x35, 0x42, 0x5b,
-	0x5f, 0xf2, 0xea, 0x3b, 0x00, 0x00, 0xff, 0xff, 0x33, 0xde, 0x2d, 0xf5, 0xf5, 0x01, 0x00, 0x00,
+	// 261 bytes of a gzipped FileDescriptorProto
+	0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0xff, 0xe2, 0xe2, 0x4e, 0xce, 0x4f, 0x49,
+	0x4d, 0xd6, 0x2b, 0x28, 0xca, 0x2f, 0xc9, 0x17, 0x62, 0x2d, 0x4f, 0x4d, 0x2c, 0x4b, 0x95, 0x12,
+	0x49, 0xcf, 0x4f, 0xcf, 0x07, 0x8b, 0xe8, 0x83, 0x58, 0x10, 0x49, 0x25, 0x25, 0x2e, 0x0e, 0xdf,
+	0xd4, 0x92, 0xc4, 0x94, 0xc4, 0x92, 0x44, 0x21, 0x31, 0x2e, 0xb6, 0xe2, 0xe4, 0x8c, 0xd4, 0xdc,
+	0x44, 0x09, 0x46, 0x05, 0x46, 0x0d, 0xde, 0x20, 0x28, 0x4f, 0x29, 0x96, 0x4b, 0x20, 0x2c, 0x31,
+	0x27, 0x33, 0x25, 0xb1, 0x24, 0xbf, 0x28, 0xb4, 0x20, 0x25, 0xb1, 0x24, 0xb5, 0x58, 0xc8, 0x93,
+	0x4b, 0xb0, 0x0c, 0x26, 0x16, 0x5f, 0x0a, 0x11, 0x94, 0x60, 0x54, 0x60, 0xd6, 0xe0, 0x36, 0x12,
+	0xd3, 0x03, 0x5b, 0xa8, 0x87, 0xa6, 0xc7, 0x89, 0xe5, 0xc4, 0x3d, 0x79, 0x86, 0x20, 0x81, 0x32,
+	0x34, 0xa3, 0x94, 0x42, 0xb9, 0xf8, 0xd1, 0x94, 0x0a, 0xe9, 0x70, 0xb1, 0x17, 0x94, 0x26, 0xc5,
+	0x67, 0xa7, 0x56, 0x82, 0x9d, 0xc2, 0x6d, 0xc4, 0x0b, 0x35, 0x33, 0xa0, 0x34, 0xc9, 0x3b, 0xb5,
+	0x12, 0x6a, 0x14, 0x5b, 0x01, 0x98, 0x27, 0x24, 0xc2, 0xc5, 0x5a, 0x90, 0x5f, 0x9e, 0x5a, 0x24,
+	0xc1, 0xa4, 0xc0, 0xa8, 0xc1, 0x1c, 0x04, 0xe1, 0x28, 0x19, 0x70, 0xb1, 0x41, 0x54, 0x0b, 0x09,
+	0x71, 0xb1, 0x94, 0x54, 0x16, 0xa4, 0x82, 0x8d, 0xe2, 0x0c, 0x02, 0xb3, 0x41, 0x62, 0x20, 0x3f,
+	0x83, 0xb5, 0xf0, 0x04, 0x81, 0xd9, 0x4e, 0x12, 0x27, 0x1e, 0xc9, 0x31, 0x5e, 0x78, 0x24, 0xc7,
+	0xf8, 0xe0, 0x91, 0x1c, 0xe3, 0x84, 0xc7, 0x72, 0x0c, 0x17, 0x1e, 0xcb, 0x31, 0xdc, 0x78, 0x2c,
+	0xc7, 0x90, 0xc4, 0x06, 0x0e, 0x2c, 0x63, 0x40, 0x00, 0x00, 0x00, 0xff, 0xff, 0x94, 0x7c, 0x13,
+	0x54, 0x58, 0x01, 0x00, 0x00,
 }
 
 func (m *Metadata) Marshal() (dAtA []byte, err error) {
@@ -397,7 +275,7 @@ func (m *Metadata) MarshalTo(dAtA []byte) (int, error) {
 	return i, nil
 }
 
-func (m *CommitInfo) Marshal() (dAtA []byte, err error) {
+func (m *ValidatorUpdates) Marshal() (dAtA []byte, err error) {
 	size := m.Size()
 	dAtA = make([]byte, size)
 	n, err := m.MarshalTo(dAtA)
@@ -407,19 +285,14 @@ func (m *CommitInfo) Marshal() (dAtA []byte, err error) {
 	return dAtA[:n], nil
 }
 
-func (m *CommitInfo) MarshalTo(dAtA []byte) (int, error) {
+func (m *ValidatorUpdates) MarshalTo(dAtA []byte) (int, error) {
 	var i int
 	_ = i
 	var l int
 	_ = l
-	if m.Round != 0 {
-		dAtA[i] = 0x8
-		i++
-		i = encodeVarintCodec(dAtA, i, uint64(m.Round))
-	}
-	if len(m.Votes) > 0 {
-		for _, msg := range m.Votes {
-			dAtA[i] = 0x12
+	if len(m.ValidatorUpdates) > 0 {
+		for _, msg := range m.ValidatorUpdates {
+			dAtA[i] = 0xa
 			i++
 			i = encodeVarintCodec(dAtA, i, uint64(msg.Size()))
 			n, err := msg.MarshalTo(dAtA[i:])
@@ -428,71 +301,6 @@ func (m *CommitInfo) MarshalTo(dAtA []byte) (int, error) {
 			}
 			i += n
 		}
-	}
-	return i, nil
-}
-
-func (m *VoteInfo) Marshal() (dAtA []byte, err error) {
-	size := m.Size()
-	dAtA = make([]byte, size)
-	n, err := m.MarshalTo(dAtA)
-	if err != nil {
-		return nil, err
-	}
-	return dAtA[:n], nil
-}
-
-func (m *VoteInfo) MarshalTo(dAtA []byte) (int, error) {
-	var i int
-	_ = i
-	var l int
-	_ = l
-	dAtA[i] = 0xa
-	i++
-	i = encodeVarintCodec(dAtA, i, uint64(m.Validator.Size()))
-	n1, err := m.Validator.MarshalTo(dAtA[i:])
-	if err != nil {
-		return 0, err
-	}
-	i += n1
-	if m.SignedLastBlock {
-		dAtA[i] = 0x10
-		i++
-		if m.SignedLastBlock {
-			dAtA[i] = 1
-		} else {
-			dAtA[i] = 0
-		}
-		i++
-	}
-	return i, nil
-}
-
-func (m *Validator) Marshal() (dAtA []byte, err error) {
-	size := m.Size()
-	dAtA = make([]byte, size)
-	n, err := m.MarshalTo(dAtA)
-	if err != nil {
-		return nil, err
-	}
-	return dAtA[:n], nil
-}
-
-func (m *Validator) MarshalTo(dAtA []byte) (int, error) {
-	var i int
-	_ = i
-	var l int
-	_ = l
-	if len(m.Address) > 0 {
-		dAtA[i] = 0xa
-		i++
-		i = encodeVarintCodec(dAtA, i, uint64(len(m.Address)))
-		i += copy(dAtA[i:], m.Address)
-	}
-	if m.Power != 0 {
-		dAtA[i] = 0x18
-		i++
-		i = encodeVarintCodec(dAtA, i, uint64(m.Power))
 	}
 	return i, nil
 }
@@ -515,11 +323,11 @@ func (m *ValidatorUpdate) MarshalTo(dAtA []byte) (int, error) {
 	dAtA[i] = 0xa
 	i++
 	i = encodeVarintCodec(dAtA, i, uint64(m.PubKey.Size()))
-	n2, err := m.PubKey.MarshalTo(dAtA[i:])
+	n1, err := m.PubKey.MarshalTo(dAtA[i:])
 	if err != nil {
 		return 0, err
 	}
-	i += n2
+	i += n1
 	if m.Power != 0 {
 		dAtA[i] = 0x10
 		i++
@@ -579,50 +387,17 @@ func (m *Metadata) Size() (n int) {
 	return n
 }
 
-func (m *CommitInfo) Size() (n int) {
+func (m *ValidatorUpdates) Size() (n int) {
 	if m == nil {
 		return 0
 	}
 	var l int
 	_ = l
-	if m.Round != 0 {
-		n += 1 + sovCodec(uint64(m.Round))
-	}
-	if len(m.Votes) > 0 {
-		for _, e := range m.Votes {
+	if len(m.ValidatorUpdates) > 0 {
+		for _, e := range m.ValidatorUpdates {
 			l = e.Size()
 			n += 1 + l + sovCodec(uint64(l))
 		}
-	}
-	return n
-}
-
-func (m *VoteInfo) Size() (n int) {
-	if m == nil {
-		return 0
-	}
-	var l int
-	_ = l
-	l = m.Validator.Size()
-	n += 1 + l + sovCodec(uint64(l))
-	if m.SignedLastBlock {
-		n += 2
-	}
-	return n
-}
-
-func (m *Validator) Size() (n int) {
-	if m == nil {
-		return 0
-	}
-	var l int
-	_ = l
-	l = len(m.Address)
-	if l > 0 {
-		n += 1 + l + sovCodec(uint64(l))
-	}
-	if m.Power != 0 {
-		n += 1 + sovCodec(uint64(m.Power))
 	}
 	return n
 }
@@ -743,7 +518,7 @@ func (m *Metadata) Unmarshal(dAtA []byte) error {
 	}
 	return nil
 }
-func (m *CommitInfo) Unmarshal(dAtA []byte) error {
+func (m *ValidatorUpdates) Unmarshal(dAtA []byte) error {
 	l := len(dAtA)
 	iNdEx := 0
 	for iNdEx < l {
@@ -766,34 +541,15 @@ func (m *CommitInfo) Unmarshal(dAtA []byte) error {
 		fieldNum := int32(wire >> 3)
 		wireType := int(wire & 0x7)
 		if wireType == 4 {
-			return fmt.Errorf("proto: CommitInfo: wiretype end group for non-group")
+			return fmt.Errorf("proto: ValidatorUpdates: wiretype end group for non-group")
 		}
 		if fieldNum <= 0 {
-			return fmt.Errorf("proto: CommitInfo: illegal tag %d (wire type %d)", fieldNum, wire)
+			return fmt.Errorf("proto: ValidatorUpdates: illegal tag %d (wire type %d)", fieldNum, wire)
 		}
 		switch fieldNum {
 		case 1:
-			if wireType != 0 {
-				return fmt.Errorf("proto: wrong wireType = %d for field Round", wireType)
-			}
-			m.Round = 0
-			for shift := uint(0); ; shift += 7 {
-				if shift >= 64 {
-					return ErrIntOverflowCodec
-				}
-				if iNdEx >= l {
-					return io.ErrUnexpectedEOF
-				}
-				b := dAtA[iNdEx]
-				iNdEx++
-				m.Round |= int32(b&0x7F) << shift
-				if b < 0x80 {
-					break
-				}
-			}
-		case 2:
 			if wireType != 2 {
-				return fmt.Errorf("proto: wrong wireType = %d for field Votes", wireType)
+				return fmt.Errorf("proto: wrong wireType = %d for field ValidatorUpdates", wireType)
 			}
 			var msglen int
 			for shift := uint(0); ; shift += 7 {
@@ -820,223 +576,11 @@ func (m *CommitInfo) Unmarshal(dAtA []byte) error {
 			if postIndex > l {
 				return io.ErrUnexpectedEOF
 			}
-			m.Votes = append(m.Votes, VoteInfo{})
-			if err := m.Votes[len(m.Votes)-1].Unmarshal(dAtA[iNdEx:postIndex]); err != nil {
+			m.ValidatorUpdates = append(m.ValidatorUpdates, ValidatorUpdate{})
+			if err := m.ValidatorUpdates[len(m.ValidatorUpdates)-1].Unmarshal(dAtA[iNdEx:postIndex]); err != nil {
 				return err
 			}
 			iNdEx = postIndex
-		default:
-			iNdEx = preIndex
-			skippy, err := skipCodec(dAtA[iNdEx:])
-			if err != nil {
-				return err
-			}
-			if skippy < 0 {
-				return ErrInvalidLengthCodec
-			}
-			if (iNdEx + skippy) < 0 {
-				return ErrInvalidLengthCodec
-			}
-			if (iNdEx + skippy) > l {
-				return io.ErrUnexpectedEOF
-			}
-			iNdEx += skippy
-		}
-	}
-
-	if iNdEx > l {
-		return io.ErrUnexpectedEOF
-	}
-	return nil
-}
-func (m *VoteInfo) Unmarshal(dAtA []byte) error {
-	l := len(dAtA)
-	iNdEx := 0
-	for iNdEx < l {
-		preIndex := iNdEx
-		var wire uint64
-		for shift := uint(0); ; shift += 7 {
-			if shift >= 64 {
-				return ErrIntOverflowCodec
-			}
-			if iNdEx >= l {
-				return io.ErrUnexpectedEOF
-			}
-			b := dAtA[iNdEx]
-			iNdEx++
-			wire |= uint64(b&0x7F) << shift
-			if b < 0x80 {
-				break
-			}
-		}
-		fieldNum := int32(wire >> 3)
-		wireType := int(wire & 0x7)
-		if wireType == 4 {
-			return fmt.Errorf("proto: VoteInfo: wiretype end group for non-group")
-		}
-		if fieldNum <= 0 {
-			return fmt.Errorf("proto: VoteInfo: illegal tag %d (wire type %d)", fieldNum, wire)
-		}
-		switch fieldNum {
-		case 1:
-			if wireType != 2 {
-				return fmt.Errorf("proto: wrong wireType = %d for field Validator", wireType)
-			}
-			var msglen int
-			for shift := uint(0); ; shift += 7 {
-				if shift >= 64 {
-					return ErrIntOverflowCodec
-				}
-				if iNdEx >= l {
-					return io.ErrUnexpectedEOF
-				}
-				b := dAtA[iNdEx]
-				iNdEx++
-				msglen |= int(b&0x7F) << shift
-				if b < 0x80 {
-					break
-				}
-			}
-			if msglen < 0 {
-				return ErrInvalidLengthCodec
-			}
-			postIndex := iNdEx + msglen
-			if postIndex < 0 {
-				return ErrInvalidLengthCodec
-			}
-			if postIndex > l {
-				return io.ErrUnexpectedEOF
-			}
-			if err := m.Validator.Unmarshal(dAtA[iNdEx:postIndex]); err != nil {
-				return err
-			}
-			iNdEx = postIndex
-		case 2:
-			if wireType != 0 {
-				return fmt.Errorf("proto: wrong wireType = %d for field SignedLastBlock", wireType)
-			}
-			var v int
-			for shift := uint(0); ; shift += 7 {
-				if shift >= 64 {
-					return ErrIntOverflowCodec
-				}
-				if iNdEx >= l {
-					return io.ErrUnexpectedEOF
-				}
-				b := dAtA[iNdEx]
-				iNdEx++
-				v |= int(b&0x7F) << shift
-				if b < 0x80 {
-					break
-				}
-			}
-			m.SignedLastBlock = bool(v != 0)
-		default:
-			iNdEx = preIndex
-			skippy, err := skipCodec(dAtA[iNdEx:])
-			if err != nil {
-				return err
-			}
-			if skippy < 0 {
-				return ErrInvalidLengthCodec
-			}
-			if (iNdEx + skippy) < 0 {
-				return ErrInvalidLengthCodec
-			}
-			if (iNdEx + skippy) > l {
-				return io.ErrUnexpectedEOF
-			}
-			iNdEx += skippy
-		}
-	}
-
-	if iNdEx > l {
-		return io.ErrUnexpectedEOF
-	}
-	return nil
-}
-func (m *Validator) Unmarshal(dAtA []byte) error {
-	l := len(dAtA)
-	iNdEx := 0
-	for iNdEx < l {
-		preIndex := iNdEx
-		var wire uint64
-		for shift := uint(0); ; shift += 7 {
-			if shift >= 64 {
-				return ErrIntOverflowCodec
-			}
-			if iNdEx >= l {
-				return io.ErrUnexpectedEOF
-			}
-			b := dAtA[iNdEx]
-			iNdEx++
-			wire |= uint64(b&0x7F) << shift
-			if b < 0x80 {
-				break
-			}
-		}
-		fieldNum := int32(wire >> 3)
-		wireType := int(wire & 0x7)
-		if wireType == 4 {
-			return fmt.Errorf("proto: Validator: wiretype end group for non-group")
-		}
-		if fieldNum <= 0 {
-			return fmt.Errorf("proto: Validator: illegal tag %d (wire type %d)", fieldNum, wire)
-		}
-		switch fieldNum {
-		case 1:
-			if wireType != 2 {
-				return fmt.Errorf("proto: wrong wireType = %d for field Address", wireType)
-			}
-			var byteLen int
-			for shift := uint(0); ; shift += 7 {
-				if shift >= 64 {
-					return ErrIntOverflowCodec
-				}
-				if iNdEx >= l {
-					return io.ErrUnexpectedEOF
-				}
-				b := dAtA[iNdEx]
-				iNdEx++
-				byteLen |= int(b&0x7F) << shift
-				if b < 0x80 {
-					break
-				}
-			}
-			if byteLen < 0 {
-				return ErrInvalidLengthCodec
-			}
-			postIndex := iNdEx + byteLen
-			if postIndex < 0 {
-				return ErrInvalidLengthCodec
-			}
-			if postIndex > l {
-				return io.ErrUnexpectedEOF
-			}
-			m.Address = append(m.Address[:0], dAtA[iNdEx:postIndex]...)
-			if m.Address == nil {
-				m.Address = []byte{}
-			}
-			iNdEx = postIndex
-		case 3:
-			if wireType != 0 {
-				return fmt.Errorf("proto: wrong wireType = %d for field Power", wireType)
-			}
-			m.Power = 0
-			for shift := uint(0); ; shift += 7 {
-				if shift >= 64 {
-					return ErrIntOverflowCodec
-				}
-				if iNdEx >= l {
-					return io.ErrUnexpectedEOF
-				}
-				b := dAtA[iNdEx]
-				iNdEx++
-				m.Power |= int64(b&0x7F) << shift
-				if b < 0x80 {
-					break
-				}
-			}
 		default:
 			iNdEx = preIndex
 			skippy, err := skipCodec(dAtA[iNdEx:])