+++ orm/sequence_test.go
@@ -33,7 +33,7 @@ func TestSequence(t *testing.T) {
 func TestSequenceKeyFormat(t *testing.T) {
 	db := store.MemStore()
 	s := NewSequence("bucket", "name")
-	_ , err := s.NextVal(db)
+	_, err := s.NextVal(db)
 	assert.Nil(t, err)
 	// As defined in NewSequence documentation
 	key := `_s.bucket:name`