+++ x/validators/init.go
@@ -6,8 +6,7 @@ import (
 )
 
 const (
-	optKey   = "update_validators"
-	storeKey = "_1:update_validators"
+	optKey = "update_validators"
 )
 
 // Initializer fulfils the InitStater interface to load data from
@@ -32,10 +31,11 @@ func (Initializer) FromGenesis(opts weave.Options, params weave.GenesisParams, k
 		return errors.Wrap(err, "bucket save")
 	}
 
-	vu := ValidatorUpdatesFromABCI(params.Validators)
+	// Deduplicate validators for storage.
+	vu := weave.ValidatorUpdatesFromABCI(params.Validators).Deduplicate(true)
 	if err := vu.Validate(); err != nil {
 		return errors.Wrap(err, "validator updates")
 	}
 
-	return errors.Wrap(vu.Store(kv), "store validator updates")
+	return errors.Wrap(weave.StoreValidatorUpdates(kv, vu), "store validator updates")
 }